@@ -14,34 +14,33 @@
     "axios": "^0.19.0",
     "core-js": "^2.6.9",
     "echarts": "^4.2.1",
-    "element-ui": "^2.11.1",
     "express": "^4.17.1",
     "js-cookie": "^2.2.1",
     "mockjs": "^1.0.1-beta3",
     "nprogress": "^0.2.0",
     "qrcodejs2": "0.0.2",
-    "vue": "^2.6.10",
-    "vue-i18n": "^8.14.0",
-    "vue-router": "^3.1.2",
-    "vuex": "^3.1.1"
+    "vue": "^3.1.1",
+    "vue-i18n": "^9.1.6",
+    "vue-router": "^4.0.8",
+    "vuex": "^4.0.1",
+    "element-plus": "^1.0.2-beta.55"
   },
   "devDependencies": {
-    "@babel/core": "^7.5.5",
+    "@babel/core": "^7.14.6",
     "@babel/plugin-transform-runtime": "^7.5.5",
     "@babel/preset-env": "^7.5.5",
-    "@vue/cli-plugin-babel": "^3.11.0",
-    "@vue/cli-plugin-eslint": "^3.11.0",
-    "@vue/cli-service": "^3.11.0",
-    "babel-eslint": "^10.0.2",
+    "@vue/cli-plugin-babel": "^4.5.0",
+    "@vue/cli-plugin-eslint": "^4.5.0",
+    "@vue/cli-service": "^4.5.0",
     "compression-webpack-plugin": "^3.0.0",
-    "eslint": "^6.2.2",
-    "eslint-plugin-vue": "^5.2.3",
+    "eslint": "^7.20.0",
+    "eslint-plugin-vue": "^7.11.1",
     "image-webpack-loader": "^5.0.0",
     "less": "^3.10.3",
     "less-loader": "^5.0.0",
     "vue-hot-reload-api": "^2.3.3",
     "vue-loader": "^15.7.1",
-    "vue-template-compiler": "^2.6.10"
+    "@vue/compiler-sfc": "^3.1.1"
   },
   "eslintConfig": {
     "root": true,
@@ -54,7 +53,7 @@
     ],
     "rules": {},
     "parserOptions": {
-      "parser": "babel-eslint"
+      "parser": "@babel/eslint-parser"
     }
   },
   "postcss": {
