@@ -16,7 +16,7 @@
         mounted(){
             this.loadChart();
         },
-		beforeDestroy() {
+		beforeUnmount() {
 			if (!this.myChart) {
 				return
 			}
