@@ -1,5 +1,7 @@
-// 存放一些全局的组件
-import Vue from 'vue'
 import IconSvg from './IconSvg'
 //全局注册icon-svg
-Vue.component('icon-svg', IconSvg)
+export default {
+  install: app => {
+    app.component('icon-svg', IconSvg);
+  }
+};;
