@@ -7,7 +7,8 @@
             ref="search_data" 
             class="demo-form-inline search-form">
             <el-form-item label="">
-                <el-input v-model="search_data.name" placeholder="用户名"  @keyup.enter.native='onScreeoutMoney("search_data")'></el-input>
+                <!-- native modifier has been removed, please confirm whether the function has been affected  -->
+                <el-input v-model="search_data.name" placeholder="用户名"  @keyup.enter='onScreeoutMoney("search_data")'></el-input>
             </el-form-item>
             <el-form-item>
                 <el-button type="primary" size ="mini" icon="search" @click='onScreeoutMoney("search_data")'>筛选</el-button>
@@ -26,7 +27,9 @@
    import { mapGetters } from 'vuex'
 
   export default {
+      emits: ["searchList", "showDialog", "onBatchDelMoney"],
       name:'searchItem',
+
       data(){
           return {
             search_data:{
@@ -41,12 +44,15 @@
             }
           }
       },
-       computed:{
-        ...mapGetters(['searchBtnDisabled']),
 
-      },
+      computed:{
+       ...mapGetters(['searchBtnDisabled']),
+
+     },
+
       created(){
       },
+
       methods:{
           onScreeoutMoney(searchForm){
               this.$refs[searchForm].validate((valid) => {
@@ -63,7 +69,7 @@
               this.$emit("onBatchDelMoney");
           }
       }
-  }
+  };
 </script>
 
 <style lang="less" scoped>
