@@ -1,5 +1,5 @@
 import axios from 'axios'
-import { Message, MessageBox } from 'element-ui'
+import { ElMessage as Message, ElMessageBox as MessageBox } from "element-plus"
 import store from '../store'
 import { getToken } from '@/utils/auth'
 
